+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:bonfire/bonfire.dart';
 import 'package:example/lpc/lpc_game.dart';
 import 'package:example/manual_map/game_manual_map.dart';
+import 'package:example/multi_scenario/multi_scenario.dart';
 import 'package:example/random_map/random_map_game.dart';
 import 'package:example/shared/enemy/goblin_controller.dart';
 import 'package:example/shared/interface/bar_life_controller.dart';
@@ -95,6 +96,14 @@ class Menu extends StatelessWidget {
                   ),
                   _buildButton(context, 'Top down game', () {
                     _navTo(context, const TopDownGame());
+                  }),
+                  const SizedBox(
+                    height: 10,
+                  ),
+                  _buildButton(context, 'Multi Scenario', () {
+                    MultiScenario.prepare().then((_) {
+                      _navTo(context, const MultiScenario());
+                    });
                   }),
                   if (!kIsWeb) ...[
                     const SizedBox(