+++ example/lib/multi_scenario/utils/enums/map_id_enum.dart
@@ -1,0 +1,1 @@
+enum MapBiomeId { none, biome1, biome2 }